Move tests for sequence function to TestSequenceFunction
