+++ presto-main/src/test/java/com/facebook/presto/type/TestArrayOperators.java
@@ -26,7 +26,6 @@ import com.facebook.presto.spi.type.ArrayType;
 import com.facebook.presto.spi.type.BooleanType;
 import com.facebook.presto.spi.type.MapType;
 import com.facebook.presto.spi.type.RowType;
-import com.facebook.presto.spi.type.SqlDate;
 import com.facebook.presto.spi.type.SqlTimestamp;
 import com.facebook.presto.spi.type.StandardTypes;
 import com.facebook.presto.spi.type.Type;
@@ -42,14 +41,9 @@ import io.airlift.slice.Slice;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.TimeZone;
-import java.util.concurrent.TimeUnit;
 
 import static com.facebook.presto.SessionTestUtils.TEST_SESSION;
 import static com.facebook.presto.block.BlockSerdeUtil.writeBlock;
@@ -61,7 +55,6 @@ import static com.facebook.presto.spi.StandardErrorCode.NOT_SUPPORTED;
 import static com.facebook.presto.spi.function.OperatorType.HASH_CODE;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
-import static com.facebook.presto.spi.type.DateType.DATE;
 import static com.facebook.presto.spi.type.DecimalType.createDecimalType;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 import static com.facebook.presto.spi.type.IntegerType.INTEGER;
@@ -1267,190 +1260,6 @@ public class TestArrayOperators
         assertInvalidFunction("REPEAT(array[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20], 9999)", INVALID_FUNCTION_ARGUMENT);
     }
 
-    @Test
-    public void testSequence()
-    {
-        // defaults to a step of 1
-        assertFunction("SEQUENCE(1, 5)", new ArrayType(BIGINT), ImmutableList.of(1L, 2L, 3L, 4L, 5L));
-        assertFunction("SEQUENCE(-10, -5)", new ArrayType(BIGINT), ImmutableList.of(-10L, -9L, -8L, -7L, -6L, -5L));
-        assertFunction("SEQUENCE(-5, 2)", new ArrayType(BIGINT), ImmutableList.of(-5L, -4L, -3L, -2L, -1L, 0L, 1L, 2L));
-        assertFunction("SEQUENCE(2, 2)", new ArrayType(BIGINT), ImmutableList.of(2L));
-
-        // defaults to a step of -1
-        assertFunction("SEQUENCE(5, 1)", new ArrayType(BIGINT), ImmutableList.of(5L, 4L, 3L, 2L, 1L));
-        assertFunction("SEQUENCE(-5, -10)", new ArrayType(BIGINT), ImmutableList.of(-5L, -6L, -7L, -8L, -9L, -10L));
-        assertFunction("SEQUENCE(2, -5)", new ArrayType(BIGINT), ImmutableList.of(2L, 1L, 0L, -1L, -2L, -3L, -4L, -5L));
-
-        // with increment
-        assertFunction("SEQUENCE(1, 9, 4)", new ArrayType(BIGINT), ImmutableList.of(1L, 5L, 9L));
-        assertFunction("SEQUENCE(-10, -5, 2)", new ArrayType(BIGINT), ImmutableList.of(-10L, -8L, -6L));
-        assertFunction("SEQUENCE(-5, 2, 3)", new ArrayType(BIGINT), ImmutableList.of(-5L, -2L, 1L));
-        assertFunction("SEQUENCE(2, 2, 2)", new ArrayType(BIGINT), ImmutableList.of(2L));
-        assertFunction("SEQUENCE(5, 1, -1)", new ArrayType(BIGINT), ImmutableList.of(5L, 4L, 3L, 2L, 1L));
-        assertFunction("SEQUENCE(10, 2, -2)", new ArrayType(BIGINT), ImmutableList.of(10L, 8L, 6L, 4L, 2L));
-
-        // failure modes
-        assertInvalidFunction("SEQUENCE(2, -1, 1)", INVALID_FUNCTION_ARGUMENT);
-        assertInvalidFunction("SEQUENCE(-1, -10, 1)", INVALID_FUNCTION_ARGUMENT);
-        assertInvalidFunction("SEQUENCE(1, 1000000)", INVALID_FUNCTION_ARGUMENT);
-    }
-
-    @Test
-    public void testSequenceDateTimeDayToSecond()
-            throws ParseException
-    {
-        assertFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-04-14', interval '1' day)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-12"), sqlDate("2016-04-13"), sqlDate("2016-04-14")));
-        assertFunction(
-                "SEQUENCE(date '2016-04-14', date '2016-04-12', interval '-1' day)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-14"), sqlDate("2016-04-13"), sqlDate("2016-04-12")));
-
-        assertFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-04-16', interval '2' day)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-12"), sqlDate("2016-04-14"), sqlDate("2016-04-16")));
-        assertFunction(
-                "SEQUENCE(date '2016-04-16', date '2016-04-12', interval '-2' day)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-16"), sqlDate("2016-04-14"), sqlDate("2016-04-12")));
-
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-16 01:07:00', interval '3' minute)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2016-04-16 01:03:10"), sqlTimestamp("2016-04-16 01:06:10")));
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:10:10', timestamp '2016-04-16 01:03:00', interval '-3' minute)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:10:10"), sqlTimestamp("2016-04-16 01:07:10"), sqlTimestamp("2016-04-16 01:04:10")));
-
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-16 01:01:00', interval '20' second)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2016-04-16 01:00:30"), sqlTimestamp("2016-04-16 01:00:50")));
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:01:10', timestamp '2016-04-16 01:00:20', interval '-20' second)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:01:10"), sqlTimestamp("2016-04-16 01:00:50"), sqlTimestamp("2016-04-16 01:00:30")));
-
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-18 01:01:00', interval '19' hour)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2016-04-16 20:00:10"), sqlTimestamp("2016-04-17 15:00:10")));
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-14 01:00:20', interval '-19' hour)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2016-04-15 06:00:10"), sqlTimestamp("2016-04-14 11:00:10")));
-
-        // failure modes
-        assertInvalidFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-04-14', interval '-1' day)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(date '2016-04-14', date '2016-04-12', interval '1' day)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(date '2000-04-14', date '2030-04-12', interval '1' day)",
-                INVALID_FUNCTION_ARGUMENT,
-                "result of sequence function must not have more than 10000 entries");
-        assertInvalidFunction(
-                "SEQUENCE(date '2018-01-01', date '2018-01-04', interval '18' hour)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence step must be a day interval if start and end values are dates");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-16 01:01:00', interval '-20' second)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:10:10', timestamp '2016-04-16 01:01:00', interval '20' second)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-04-16 09:01:00', interval '1' second)",
-                INVALID_FUNCTION_ARGUMENT,
-                "result of sequence function must not have more than 10000 entries");
-    }
-
-    @Test
-    public void testSequenceDateTimeYearToMonth()
-            throws ParseException
-    {
-        assertFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-06-12', interval '1' month)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-12"), sqlDate("2016-05-12"), sqlDate("2016-06-12")));
-        assertFunction(
-                "SEQUENCE(date '2016-06-12', date '2016-04-12', interval '-1' month)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-06-12"), sqlDate("2016-05-12"), sqlDate("2016-04-12")));
-
-        assertFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-08-12', interval '2' month)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-12"), sqlDate("2016-06-12"), sqlDate("2016-08-12")));
-        assertFunction(
-                "SEQUENCE(date '2016-08-12', date '2016-04-12', interval '-2' month)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-08-12"), sqlDate("2016-06-12"), sqlDate("2016-04-12")));
-
-        assertFunction(
-                "SEQUENCE(date '2016-04-12', date '2018-04-12', interval '1' year)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2016-04-12"), sqlDate("2017-04-12"), sqlDate("2018-04-12")));
-        assertFunction(
-                "SEQUENCE(date '2018-04-12', date '2016-04-12', interval '-1' year)",
-                new ArrayType(DATE),
-                ImmutableList.of(sqlDate("2018-04-12"), sqlDate("2017-04-12"), sqlDate("2016-04-12")));
-
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2016-09-16 01:10:00', interval '2' month)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2016-06-16 01:00:10"), sqlTimestamp("2016-08-16 01:00:10")));
-        assertFunction(
-                "SEQUENCE(timestamp '2016-09-16 01:10:10', timestamp '2016-04-16 01:00:00', interval '-2' month)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-09-16 01:10:10"), sqlTimestamp("2016-07-16 01:10:10"), sqlTimestamp("2016-05-16 01:10:10")));
-
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '2021-04-16 01:01:00', interval '2' year)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:00:10"), sqlTimestamp("2018-04-16 01:00:10"), sqlTimestamp("2020-04-16 01:00:10")));
-        assertFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:01:10', timestamp '2011-04-16 01:00:00', interval '-2' year)",
-                new ArrayType(TIMESTAMP),
-                ImmutableList.of(sqlTimestamp("2016-04-16 01:01:10"), sqlTimestamp("2014-04-16 01:01:10"), sqlTimestamp("2012-04-16 01:01:10")));
-
-        // failure modes
-        assertInvalidFunction(
-                "SEQUENCE(date '2016-06-12', date '2016-04-12', interval '1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(date '2016-04-12', date '2016-06-12', interval '-1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(date '2000-04-12', date '3000-06-12', interval '1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "result of sequence function must not have more than 10000 entries");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-05-16 01:00:10', timestamp '2016-04-16 01:01:00', interval '1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:10:10', timestamp '2016-05-16 01:01:00', interval '-1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "sequence stop value should be greater than or equal to start value if step is greater than zero otherwise stop should be less than or equal to start");
-        assertInvalidFunction(
-                "SEQUENCE(timestamp '2016-04-16 01:00:10', timestamp '3000-04-16 09:01:00', interval '1' month)",
-                INVALID_FUNCTION_ARGUMENT,
-                "result of sequence function must not have more than 10000 entries");
-    }
-
     @Override
     public void assertInvalidFunction(String projection, SemanticErrorCode errorCode)
     {
@@ -1544,21 +1353,5 @@ public class TestArrayOperators
     private static SqlTimestamp sqlTimestamp(long millisUtc)
     {
         return new SqlTimestamp(millisUtc, TEST_SESSION.getTimeZoneKey());
-    }
-
-    private static SqlTimestamp sqlTimestamp(String dateString)
-            throws ParseException
-    {
-        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-        return sqlTimestamp(dateFormat.parse(dateString).getTime());
-    }
-
-    private static SqlDate sqlDate(String dateString)
-            throws ParseException
-    {
-        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-        return new SqlDate(toIntExact(TimeUnit.MILLISECONDS.toDays(dateFormat.parse(dateString).getTime())));
     }
 }
